// src/entities/user/model/privileges.ts
export enum UserPrivileges {
  NOTIFICATIONS = 'notifications',
  REPORTS = 'reports',
  STATISTICS = 'statistics',
  PROVIDER_TYPES = 'provider_types',
  PROVIDERS = 'providers',
  ARTIFACTS = 'artifacts',
  RESOURCE_TEMPLATES = 'resource_templates',
  GROUPS = 'groups',
  STANDS = 'stands',
  STAND_CATEGORIES = 'stand_categories',
  USERS = 'users',
  ROLES = 'roles',
  CLUSTERS_AND_NAMESPACES = 'clusters_and_namespaces',
  VMPROFILES = 'vmprofiles',
  SOFTWARE = 'software',
  MAINTENANCE = 'maintenance',
}